SELC-2803 propagazione 502 in checkResponseErrorCode, refactor ResourceNotFoundException
