+++ connector-api/src/main/java/it/pagopa/selfcare/party/registry_proxy/connector/exception/ResourceNotFoundException.java
@@ -1,6 +1,9 @@
 package it.pagopa.selfcare.party.registry_proxy.connector.exception;
 
 public class ResourceNotFoundException extends RuntimeException {
+    public ResourceNotFoundException() {
+    }
+
     public ResourceNotFoundException(String message) {
         super(message);
     }
