--- core/src/main/java/it/pagopa/selfcare/party/registry_proxy/core/exception/ResourceNotFoundException.java
@@ -1,10 +1,0 @@
-package it.pagopa.selfcare.party.registry_proxy.core.exception;
-
-public class ResourceNotFoundException extends RuntimeException {
-    public ResourceNotFoundException() {
-    }
-
-    public ResourceNotFoundException(String message) {
-        super(message);
-    }
-}
